refacytoring use page in separate omponent
